rm long()

+++ mcomix/thumbnail_tools.py
@@ -244,7 +244,7 @@ class Thumbnailer(object):
                     return False
 
                 info = img.info
-                stored_mtime = long(info['Thumb::MTime'])
+                stored_mtime = info['Thumb::MTime']
                 # The source file might no longer exist
                 file_mtime = os.path.isfile(filepath) and long(os.stat(filepath).st_mtime) or stored_mtime
                 return stored_mtime == file_mtime and \
